Adjust namespaces #39
