+++ src/Command/ListenCommand.php
@@ -2,28 +2,30 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Queue\Commands;
+namespace Yiisoft\Yii\Queue\Command;
 
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Yiisoft\Yii\Queue\Queue;
 
-class RunCommand extends Command
+class ListenCommand extends Command
 {
-    protected static $defaultName = 'queue/run';
+    protected static $defaultName = 'queue/listen';
+    /**
+     * @var Queue
+     */
     private Queue $queue;
 
     public function __construct(?string $name, Queue $queue)
     {
         parent::__construct($name);
         $this->queue = $queue;
-
-        $this->setDescription('Runs all the existing messages in the queue. Exits once messages are over.');
+        $this->setDescription('Listens the queue and executes messages as they come. Needs to be stopped manually.');
     }
 
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $this->queue->run();
+        $this->queue->listen();
     }
 }
