+++ src/Event/BeforeExecution.php
@@ -2,26 +2,26 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Queue\Events;
+namespace Yiisoft\Yii\Queue\Event;
 
-use Yiisoft\Yii\Queue\Jobs\JobInterface;
+use Yiisoft\Yii\Queue\MessageInterface;
 use Yiisoft\Yii\Queue\Queue;
 
-final class BeforePush
+final class BeforeExecution
 {
     private bool $stop = false;
     private Queue $queue;
-    private JobInterface $job;
+    private MessageInterface $message;
 
-    public function __construct(Queue $queue, JobInterface $job)
+    public function __construct(Queue $queue, MessageInterface $message)
     {
         $this->queue = $queue;
-        $this->job = $job;
+        $this->message = $message;
     }
 
-    public function getJob(): JobInterface
+    public function getMessage(): MessageInterface
     {
-        return $this->job;
+        return $this->message;
     }
 
     public function getQueue(): Queue
