+++ src/Job/CallableJob.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Yii\Queue\Jobs;
+namespace Yiisoft\Yii\Queue\Job;
 
 use Yiisoft\Serializer\SerializerInterface;
